+++ dart/dynamics/MeshShape.hpp
@@ -70,7 +70,7 @@ public:
   virtual ~MeshShape();
 
   // Documentation inherited.
-  const std::string& getType() const;
+  const std::string& getType() const override;
 
   /// Returns shape type for this class
   static const std::string& getStaticType();