+++ dart/dynamics/PlaneShape.cpp
@@ -45,7 +45,7 @@ PlaneShape::PlaneShape(const Eigen::Vector3d& _normal, double _offset)
 //==============================================================================
 PlaneShape::PlaneShape(const Eigen::Vector3d& _normal,
                        const Eigen::Vector3d& _point)
-  : Shape(PLANE),
+  : Shape(),
     mNormal(_normal.normalized()),
     mOffset(mNormal.dot(_point))
 {