+++ dart/gui/OpenGLRenderInterface.hpp
@@ -84,6 +84,8 @@ public:
     void drawEllipsoid(const Eigen::Vector3d& _size) override;
     void drawCube(const Eigen::Vector3d& _size) override;
     void drawCylinder(double _radius, double _height) override;
+    void drawCapsule(double radius, double height) override;
+    void drawCone(double radius, double height) override;
     void drawMesh(const Eigen::Vector3d& _scale, const aiScene* _mesh) override;
     void drawSoftMesh(const aiMesh* mesh) override;
     void drawList(GLuint index) override;