+++ dart/simulation/World.hpp
@@ -49,7 +49,7 @@
 #include "dart/common/Subject.hpp"
 #include "dart/dynamics/SimpleFrame.hpp"
 #include "dart/dynamics/Skeleton.hpp"
-#include "dart/collision/Option.hpp"
+#include "dart/collision/CollisionOption.hpp"
 #include "dart/simulation/Recording.hpp"
 
 namespace dart {