+++ dart/utils/sdf/SdfParser.cpp
@@ -167,6 +167,11 @@ dynamics::ShapeNode* readShapeNode(
     const std::string& skelPath,
     const common::ResourceRetrieverPtr& retriever);
 
+
+void readMaterial(
+    tinyxml2::XMLElement* materialEle,
+    dynamics::ShapeNode* shapeNode);
+
 void readVisualizationShapeNode(
     dynamics::BodyNode* bodyNode,
     tinyxml2::XMLElement* vizShapeNodeEle,
@@ -634,6 +639,10 @@ std::pair<dynamics::Joint*,dynamics::BodyNode*> createJointAndNodePair(
     return skeleton->createJointAndBodyNodePair<dynamics::BallJoint>(parent,
           static_cast<const dynamics::BallJoint::Properties&>(*joint.properties),
           static_cast<const typename NodeType::Properties&>(*node.properties));
+  else if (std::string("fixed") == type)
+    return skeleton->createJointAndBodyNodePair<dynamics::WeldJoint>(parent,
+          static_cast<const dynamics::WeldJoint::Properties&>(*joint.properties),
+          static_cast<const typename NodeType::Properties&>(*node.properties));
   else if (std::string("free") == type)
     return skeleton->createJointAndBodyNodePair<dynamics::FreeJoint>(parent,
           static_cast<const dynamics::FreeJoint::Properties&>(*joint.properties),
@@ -981,6 +990,29 @@ dynamics::ShapeNode* readShapeNode(
   return shapeNode;
 }
 
+
+//==============================================================================
+void readMaterial(
+  tinyxml2::XMLElement* materialEle,
+  dynamics::ShapeNode* shapeNode) {
+
+  auto visualAspect = shapeNode->getVisualAspect();
+  if (hasElement(materialEle, "diffuse")) {
+    Eigen::VectorXd color = getValueVectorXd(materialEle, "diffuse");
+    if (color.size() == 3) {
+      Eigen::Vector3d color3d = color;
+      visualAspect->setColor(color3d);
+    } else if (color.size() == 4) {
+      Eigen::Vector4d color4d = color;
+      visualAspect->setColor(color4d);
+    } else {
+      dterr << "[SdfParse::readMaterial] Unsupported color vector size: "
+            << color.size() << "\n";
+    }
+  }
+}
+
+
 //==============================================================================
 void readVisualizationShapeNode(
     dynamics::BodyNode* bodyNode,
@@ -994,6 +1026,13 @@ void readVisualizationShapeNode(
                       skelPath, retriever);
 
   newShapeNode->createVisualAspect();
+
+  // Material
+  if (hasElement(vizShapeNodeEle, "material"))
+  {
+    tinyxml2::XMLElement* materialEle = getElement(vizShapeNodeEle, "material");
+    readMaterial(materialEle, newShapeNode);
+  }
 }
 
 //==============================================================================